Add Lua bridge module and streamline DLL entry
