+++ UOWalkPatch/include/Engine/LuaBridge.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace Engine::Lua {
+    bool InitLuaBridge();
+    void ShutdownLuaBridge();
+    void RegisterOurLuaFunctions();
+}
+
+extern "C" __declspec(dllexport) void __stdcall SendRaw(const void* bytes, int len);
