+++ UOWalkPatch/include/Net/PacketTrace.hpp
@@ -1,0 +1,7 @@
+#pragma once
+
+namespace Net {
+    bool InitPacketTrace();
+    void ShutdownPacketTrace();
+}
+
