+++ UOWalkPatch/include/Net/SendBuilder.hpp
@@ -1,0 +1,11 @@
+#pragma once
+#include <cstdint>
+#include "Engine/GlobalState.hpp"
+
+namespace Net {
+    bool InitSendBuilder(GlobalStateInfo* state);
+    void ShutdownSendBuilder();
+    bool SendPacketRaw(const void* bytes, int len);
+    bool IsSendReady();
+}
+
